==> bin/1_clean_from_source.ipynb <==
import pandas as pd
import numpy as np

# load data
data = pd.read_csv('..//data//sf_open//311_cases.csv')

# select human waster
data_use = data[data['Request Type']=='Human Waste']
data_use = data_use[data_use['Longitude']>-140]
data_use = data_use[data_use['Latitude']!=0]

# convert datetime to rounded datetime
data_use['Opened'] = pd.to_datetime(data['Opened'])
data_use['Opened_rnd'] = data_use['Opened'].apply(lambda x: x.replace(microsecond=0,second=0,minute=0))

# export to csv
data_use.to_csv('..//data//created//311_cases_clean.csv')
